Exponential memory allocation in BeadEventFileWriter fixed. Stringbuilder with big data were problematic. Quality of life code changes

+++ Application/FileIO/PlateStatusFileWriter.cs
@@ -5,6 +5,7 @@ namespace DIOS.Application.FileIO
 {
   public class PlateStatusFileWriter
   {
+    private const string STATUSFILENAME = "StatusFile.json";
     private ResultsPublisher _publisher;
     public PlateStatusFileWriter(ResultsPublisher publisher)
     {
@@ -13,10 +14,13 @@ namespace DIOS.Application.FileIO
 
     public void Overwrite(string plateData)
     {
+      var directoryName = Path.Combine(_publisher.Outdir, ResultsPublisher.STATUSFOLDERNAME);
+      if (!_publisher.OutputDirectoryExists(directoryName))
+        return;
       //overwrite the whole thing
       try
       {
-        var fullName = Path.Combine(_publisher.Outdir, ResultsPublisher.STATUSFOLDERNAME, "StatusFile.json");
+        var fullName = Path.Combine(_publisher.Outdir, ResultsPublisher.STATUSFOLDERNAME, STATUSFILENAME);
         File.WriteAllText(fullName, plateData);
       }
       catch (Exception e)
